bug fix for bessel function calculations. on maverick converting to double was causing cuda errors, so reverted back to floats.
git-svn-id: svn://gk.umd.edu/noah/branches/gryfx_march15@2686 28a56f14-469b-1442-acfb-e68ebda8abd8

+++ device_funcs.cu
@@ -16,11 +16,11 @@ __device__ int iget_idy(void) {return blockIdx.y*blockDim.y+threadIdx.y;}
 __device__ int iget_idz(void) {return blockIdx.z*blockDim.z+threadIdx.z;}
 */
 
-__device__ double g0(double b) {
+__device__ float g0(float b) {
 
-  double tol = 1.e-7;
-  double tk, b2, b2sq;
-  double g, x, xi, err;
+  float tol = 1.e-7;
+  float tk, b2, b2sq;
+  float g, x, xi, err;
 
   if (b < tol) {return 1.0;}
 
@@ -45,11 +45,11 @@ __device__ double g0(double b) {
 
 }
 
-__device__ double g1(double b) {
+__device__ float g1(float b) {
 
-  double tol = 1.e-7;
-  double tk, b2, b2sq;
-  double g, x, xi, xp1i, err;
+  float tol = 1.e-7;
+  float tk, b2, b2sq;
+  float g, x, xi, xp1i, err;
 
   if (b < tol) {return 0.0;}
 
@@ -75,21 +75,21 @@ __device__ double g1(double b) {
 
 }
 
-__device__ double sgam0 (double b) {return sqrt(g0(b));}
+__device__ float sgam0 (float b) {return sqrt(g0(b));}
 
-__device__ double flr(double b) {
-  double gam0 = g0(b);
-  double gam1 = g1(b);
-  double gam = sqrt(gam0);
-  double rg = gam1/gam0;
+__device__ float flr(float b) {
+  float gam0 = g0(b);
+  float gam1 = g1(b);
+  float gam = sqrt(gam0);
+  float rg = gam1/gam0;
   return -b/2. * (1.-rg) * gam;
 }
 
-__device__ double flr2(double b) {
-  double gam0 = g0(b);
-  double gam1 = g1(b);
-  double gam = sqrt(gam0);
-  double rg = gam1/gam0;
+__device__ float flr2(float b) {
+  float gam0 = g0(b);
+  float gam1 = g1(b);
+  float gam = sqrt(gam0);
+  float rg = gam1/gam0;
   return -(b/2.) * (1.+(1.-rg)*(1.-(b/2.)*(3.+rg))) * gam;
 }
